backend/agentic-rag/requirements.txt
# Core FastAPI & Web Framework (skip if already installed by main backend)
# fastapi>=0.115.0
# uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0

# LangChain Core (use latest compatible versions)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# LangGraph for Multi-Agent Orchestration
# Note: langgraph-supervisor requires langgraph<0.7.0
langgraph>=0.6.0,<0.7.0
langgraph-supervisor>=0.0.29

# Google Gemini Integration
langchain-google-genai>=3.0.0

# Vector Store (FAISS for local development)
faiss-cpu>=1.8.0

# Web Search Tool
tavily-python>=0.5.0

# Document Processing
pypdf>=5.0.0
python-docx>=1.1.0
# python-multipart already in main requirements

# Text Processing
tiktoken>=0.8.0

# Utilities
aiofiles>=24.1.0
httpx>=0.27.0

# Streaming & Persistence
sse-starlette>=1.8.0
langgraph-checkpoint-sqlite>=2.0.0
langgraph-checkpoint-postgres>=2.0.0
aiosqlite>=0.20.0
psycopg[binary,pool]>=3.2.0
